requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
google-genai
google-cloud-storage
google-auth
python-multipart
pydantic-settings
firebase-admin
upstash-redis
google-cloud-firestore
google-cloud-pubsub